Install needed packages for Debian/Raspbian
Most Linux system need additional packages to be installed,
X11 and Motiv.
Basic instructions for RHEL and Ubuntu can be found here:
https://epics.anl.gov/extensions/medm/index.php

However, Debian is not mentioned.
Beside that. it is more comfortable to let the "make" command do the job.
Smuggle in an automatic install.
It needs /epicsextfinddist.sh from extensions.

Today it is tested on Debian and Raspbian, that is what I have.
More work needs to be done for RHEL, Ubuntu and other systems,
but they are better done in an own commit.